Enhanced Dialog systems for compatability with Quests. Created new prefabs for the dialog system. Created prefab templates for testing
Added in provisions in the dialog system for quests to be assigned to players at the end of certain conversations. Hence variable has been added in the system to track and make calls tho the players quest list.
New prefabs have been created for the system including new templates which will be used for conversations and testing. Modified quest allocation system and quest check systems in relation to checking objectives for ietm quests.