Updated in-dialog quest allocation, added provisions for testing systems
Quests are now allocatable variables within the dialog system rather than components. Utility functions have also been added to the dialog system to assign a quest to the player if it has been detected. The prefabs for the object have been updated accordingly.
Routienes have also been added to allow for testing of the quest and dialog systems.