Prefab modifications for system
Modified components of the NPC prefab in order to register collisions. Fixed volume collision bug.
Modified test script to have the collision event handler recognise a player entity and stop the player from moving. (TODO - Add in routienes to allow the player to talk with the NPC and test the dialog system)