Fix dialog system bugs. Created test conversation and a test quest
Several dialog system bugs were discovered.
There was a mild bug with the dialog system in relation to the system not finishing conversations and allowing the player to resume the game. This has now been fixed.
There was also a bug involving the missionToAssign variable not being read for a check when the variable was null which has now been fixed.
A bug within the input system which ment input was detected multiple times was also fixed.
The Dialog system now functions to the game specification with the NPCs able to have conversations with the players according to a dialog tree specified by a designer.
A test conversation and quest was also created hence the addition of several prefabs.
(TODO - test the quest system and consider system optimisations as well as implimenting UI interactions to print messages on screen).