Quest system complete and fixed
The quest system has now been fully implimented correctly allowing for quests be created and for players to retrieve and complete quests. The system uses and event tags and event classification values which are used to find and update perticular objectives. Upon an action being compleated a call can be made to the event handaling system whereby all objectives in the active quest are checked and updated.
Bugs with the system in relation to objective state updates have also been fixed.
(TODO - Add in routienes to process items in an inventory when an objective which involves item collection occurs).