Add files via upload
Notes : 

The script "player_animation_script" is used to check the movement direction of the object it's attached to and tell the animator what direction the character / NPC is moving. 

Currently there are only the Panda, Demon and Human sprites animated.

There is a prefab for each the Panda, Demon and Human types.