Modified player input system
The players input system has been expanded and will now retrieve input from either the keyboard or the virtual joystick depending on what platform the app is built for. PC builds read from the keyboard and ignore the virtual joystick and visa versa for mobile builds of the game. Ajustments were also made to various input settings such as sensitivity to accomodate for the change.