Improved branching in quest system and modified spawn with quest script to allow for the player to spawn with a chodsen quest.
Made slight changes to the objective status check routiene where upon an objectives completion the number of objectives to be loaded is then checked. If this evaluates to 0 then the quest will be marked as compleate and the check will be suspended allowing for quests to branch. Also a change was made to the SpawnWithQuest script to allow the quest that the user wants to assign to the player to be chosen via the variable x.