Password Minigame gameplay components inplimented and additions made to the UI.
All key function components for the minigame have been implimented. Allowing the user to select phrases to add to the password and then check the password. A new data structure was added (passphrase) which is used to assist with the scoring of the password. Two new UI elements were also added to the system, a panel which will list the phrase options and submit button which is intended to score the users password when clicked.