Added basic minigame exit routiene. Updated PasswordMinigameWindow.cs script and fixed critical bug.
New routienes were added to the evaluatePassword routiene to update the quest state when the user sucsessfuly completes a minigame. Upon doing so the player should not have to set a password for that terminal again. Hence I hhave changed the variable in the minigame activation script to check whether or not the game for that 'terminal' has been compleated. Alongside these changes a critical bug which was causing the objective compleated value not to start at 0 (Resulting in Array.OutOfRangeExemption) was fixed.