Added in light procedral generation for password quest and added new phrases for the user to choose from.
A new routiene was added to choose the prases that the user is presented with. The routiene will choose phrases at random and ensure there is a valid password this way preventing the occurance of a fail state. In addition the Error messasge box was updated to why the password was wrong to the end user.