Finalised duplicate checker for password minigame and modified save system.
The password minigame now sucsessfuly checks for duplicates and subsequently removes them. This was done using a simple iterative check after reviewing the previous implimentation. Alongside this an entry was added to prevent player progress from being saved during testing however this should be active during the release build of the game.